drop omxplayer
