+++ addons/resource.language.en_gb/resources/strings.po
@@ -7255,16 +7255,7 @@ msgctxt "#13457"
 msgid "Prefer VAAPI render method"
 msgstr ""
 
-#: system/settings/settings.xml
-msgctxt "#13458"
-msgid "Allow hardware acceleration - OMXPlayer"
-msgstr ""
-
-#. Description of setting with label #13457 "Allow hardware acceleration (OMXPlayer)"
-#: system/settings/settings.xml
-msgctxt "#13459"
-msgid "Use OMXPlayer for decoding of video files."
-msgstr ""
+#empty strings from id 13458 to 13459
 
 #: system/settings/settings.xml
 msgctxt "#13460"
