+++ system/settings/rbp2.xml
@@ -1,14 +1,5 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <settings version="1">
-  <section id="player">
-    <category id="videoplayer">
-      <group id="3">
-        <setting id="videoplayer.useomxplayer">
-          <default>false</default>
-        </setting>
-      </group>
-    </category>
-  </section>
   <section id="system">
     <category id="display">
       <group id="1">
