+++ xbmc/cores/omxplayer/OMXImage.h
@@ -12,7 +12,6 @@
 
 #include <IL/OMX_Video.h>
 
-#include "OMXClock.h"
 #if defined(STANDALONE)
 #define XB_FMT_A8R8G8B8 1
 #include "File.h"
