--- xbmc/cores/omxplayer/omxplayer_advancedsettings.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<advancedsettings>
-  <video>
-    <defaultplayer>VideoPlayer</defaultplayer>
-    <defaultVideoPlayer>VideoPlayer</defaultVideoPlayer>
-  </video>
-</advancedsettings>
