+++ xbmc/platform/linux/OMXCore.cpp
@@ -12,7 +12,6 @@
 #include "OMXCore.h"
 #include "utils/log.h"
 
-#include "OMXClock.h"
 #include "platform/linux/RBP.h"
 #include <cassert>
