+++ xbmc/settings/SettingConditions.cpp
@@ -290,9 +290,6 @@ void CSettingConditions::Initialize()
 #ifdef HAS_ZEROCONF
   m_simpleConditions.insert("has_zeroconf");
 #endif
-#ifdef TARGET_RASPBERRY_PI
-  m_simpleConditions.insert("has_omxplayer");
-#endif
 #ifdef HAVE_LIBVA
   m_simpleConditions.insert("have_libva");
 #endif
